+++ src/components/form-error/__tests__/form-errors.component.test.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import FormErrors from '../form_errors.component';
+import FormErrors from '../form-errors.component';
 
 describe('FormErros Component', () => {
   describe('when just one error', () => {
