+++ src/pages/companies/company-list.component.js
@@ -3,17 +3,13 @@ import CompaniesListing from '../../components/companies-list/companies-list';
 import Template from '../../components/template/template.component';
 import { getCompanies } from '../../service/company';
 import { toast } from 'react-toastify';
-import {
-  CircularProgress,
-  Box,
-  Button,
-  Typography,
-  Paper,
-} from '@material-ui/core';
+import { CircularProgress, Box, Button } from '@material-ui/core';
+import { useHistory } from 'react-router-dom';
 
 const CompaniesList = () => {
   const [companies, setCompanies] = useState([]);
   const [loading, setLoading] = useState(false);
+  const history = useHistory();
 
   const onStart = () => {
     setLoading(true);
@@ -52,7 +48,11 @@ const CompaniesList = () => {
       <Template
         title='Empresas'
         rightActions={
-          <Button variant='contained' color='primary' href='/companies/new'>
+          <Button
+            variant='contained'
+            color='primary'
+            onClick={() => history.push('/companies/new')}
+          >
             Cadastrar Empresa
           </Button>
         }
