+++ src/routes/index.js
@@ -3,9 +3,9 @@ import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import SecureRoute from './secure-route.component';
 import LoginPage from '../pages/login/login.component';
 import Dashboard from '../pages/dashboard/dashboard.component';
-import userList from '../pages/user-list/user-list.component';
-import CompaniesList from '../pages/companies-list/companies-list';
-import ComapanyRegister from '../pages/companies/company_register.component';
+import UserList from '../pages/users/user-list.component';
+import CompaniesList from '../pages/companies/company-list.component';
+import CompanyRegister from '../pages/companies/company-register.component';
 
 const Routes = () => {
   return (
@@ -13,12 +13,12 @@ const Routes = () => {
       <Switch>
         <Route path='/' exact component={LoginPage} />
         <SecureRoute path='/dashboard' component={Dashboard} />
-        <SecureRoute path='/users' component={userList} admin />
+        <SecureRoute path='/users' component={UserList} admin />
         <SecureRoute path='/companies' exact component={CompaniesList} admin />
         <SecureRoute
           path='/companies/new'
           exact
-          component={ComapanyRegister}
+          component={CompanyRegister}
           admin
         />
       </Switch>
